requirements.txt
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
exchange_calendars==4.6
filelock==3.16.1
Flask==3.1.0
fonttools==4.54.1
frozendict==2.4.6
fsspec==2024.9.0
html5lib==1.1
huggingface-hub==0.26.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
korean-lunar-calendar==0.3.1
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
multitasking==0.0.11
networkx==3.4.1
nltk==3.9.1
numpy==2.1.2
pandas==2.2.3
pandas_market_calendars==4.4.2
peewee==3.17.7
pillow==11.0.0
praw==7.8.1
prawcore==2.4.0
pyluach==2.2.0
pyparsing==3.2.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
soupsieve==2.6
sympy==1.13.1
ta==0.11.0
threadpoolctl==3.5.0
tokenizers==0.20.1
toolz==1.0.0
torch==2.5.0
tqdm==4.66.5
transformers==4.45.2
tzdata==2024.2
update-checker==0.18.0
urllib3==2.2.3
vaderSentiment==3.3.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
yfinance==0.2.44
